Refactor linker and implement the topology lock
