+++ lib/wasix/tests/runners.rs
@@ -140,107 +140,6 @@ mod wasi {
     }
 }
 
-#[cfg(feature = "webc_runner_rt_wcgi")]
-mod wcgi {
-    use std::{future::Future, sync::Arc};
-
-    use futures::{SinkExt, StreamExt, channel::mpsc::Sender, future::AbortHandle};
-    use rand::RngExt;
-    use tokio::runtime::Handle;
-    use wasmer_package::utils::from_bytes;
-    use wasmer_wasix::{
-        bin_factory::BinaryPackage,
-        runners::wcgi::{NoOpWcgiCallbacks, WcgiRunner},
-    };
-
-    use super::*;
-
-    #[tokio::test]
-    async fn can_run_staticserver() {
-        let webc = download_cached("https://wasmer.io/Michael-F-Bryan/staticserver@1.0.3").await;
-        let container = from_bytes(webc).unwrap();
-
-        let entrypoint = container.manifest().entrypoint.as_ref().unwrap();
-        assert!(WcgiRunner::can_run_command(&container.manifest().commands[entrypoint]).unwrap());
-    }
-
-    #[tokio::test]
-    async fn staticserver() {
-        let webc = download_cached("https://wasmer.io/Michael-F-Bryan/staticserver@1.0.3").await;
-        let (rt, tasks) = runtime();
-        let container = from_bytes(webc).unwrap();
-        let mut runner = WcgiRunner::new(NoOpWcgiCallbacks);
-        let port = rand::rng().random_range(10000_u16..65535_u16);
-        let (cb, started) = callbacks(Handle::current());
-        runner
-            .config()
-            .addr(([127, 0, 0, 1], port).into())
-            .callbacks(cb);
-        let pkg = BinaryPackage::from_webc(&container, &rt).await.unwrap();
-
-        // The server blocks so we need to start it on a background thread.
-        let join_handle = std::thread::spawn(move || {
-            let _guard = tasks.runtime_handle().enter();
-            runner.run_command("serve", &pkg, Arc::new(rt)).unwrap();
-        });
-
-        // wait for the server to have started
-        let abort_handle = started.await;
-
-        // Now the server is running, we can check that it is working by
-        // fetching "/" and checking for known content. We also want the server
-        // to close connections immediately so the graceful shutdown can kill
-        // the server immediately instead of waiting for the connection to time
-        // out.
-        let resp = client()
-            .get(format!("http://localhost:{port}/"))
-            .header("Connection", "close")
-            .send()
-            .await
-            .unwrap();
-        let body = resp.error_for_status().unwrap().text().await.unwrap();
-        assert!(body.contains("<title>Index of /</title>"), "{}", body);
-
-        // Make sure the server is shutdown afterwards. Failing tests will leak
-        // the server thread, but that's fine.
-        abort_handle.abort();
-        if let Err(e) = join_handle.join() {
-            std::panic::resume_unwind(e);
-        }
-    }
-
-    fn callbacks(handle: Handle) -> (Callbacks, impl Future<Output = AbortHandle>) {
-        let (sender, mut rx) = futures::channel::mpsc::channel(1);
-
-        let cb = Callbacks { sender, handle };
-        let fut = async move { rx.next().await.unwrap() };
-
-        (cb, fut)
-    }
-
-    #[derive(Debug)]
-    struct Callbacks {
-        sender: Sender<AbortHandle>,
-        handle: Handle,
-    }
-
-    impl wasmer_wasix::runners::wcgi::Callbacks for Callbacks {
-        fn started(&self, abort: futures::stream::AbortHandle) {
-            let mut sender = self.sender.clone();
-            self.handle.spawn(async move {
-                sender.send(abort).await.unwrap();
-            });
-        }
-
-        fn on_stderr(&self, stderr: &[u8]) {
-            panic!(
-                "Something was written to stderr: {}",
-                String::from_utf8_lossy(stderr)
-            );
-        }
-    }
-}
-
 async fn download_cached(url: &str) -> bytes::Bytes {
     let uri: http::Uri = url.parse().unwrap();
