+++ lib/wasix/tests/wasm_tests/dynamic_library_tests/dl-multi-thread-shared-side-calls/side.c
@@ -1,0 +1,1 @@
+int side_func(int x) { return x * 2; }
